+++ app/lib/shared/versions.dart
@@ -22,7 +22,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2022.02.16',
+  '2022.02.21',
   // Fallback runtime versions.
   '2022.02.09',
   '2022.01.31',
@@ -55,9 +55,9 @@ bool shouldGCVersion(String version) =>
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.16.1';
 final String toolStableDartSdkVersion = '2.16.1';
-final String toolStableFlutterSdkVersion = '2.10.1';
+final String toolStableFlutterSdkVersion = '2.10.2';
 final String toolPreviewDartSdkVersion = '2.17.0-85.0.dev';
-final String toolPreviewFlutterSdkVersion = '2.10.1';
+final String toolPreviewFlutterSdkVersion = '2.11.0-0.1.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);