+++ CHANGELOG.md
@@ -2,11 +2,11 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bumped runtimeVersion to `2022.02.16`.
+ * Bumped runtimeVersion to `2022.02.21`.
  * Upgraded runtime Dart SDK to `2.16.1`
  * Upgraded preview Dart analysis SDK to `2.17.0-85.0.dev`.
- * Upgraded stable Flutter analysis SDK to `2.10.1`.
- * Upgraded preview Flutter analysis SDK to `2.10.1`.
+ * Upgraded stable Flutter analysis SDK to `2.10.2`.
+ * Upgraded preview Flutter analysis SDK to `2.11.0-0.1.pre`.
  * Upgraded dartdoc to `5.0.0`.
 
 ## `20220209t173000-all`